refactor: move client runtime sources to embedded js
